backend/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
pydantic==2.10.4
pydantic-settings==2.7.0
python-multipart==0.0.20
passlib[bcrypt]==1.7.4
python-slugify==8.0.4
opencv-contrib-python-headless==4.10.0.84
numpy==2.2.0
Pillow==11.0.0
scikit-image==0.24.0
PyMuPDF==1.25.1
pytesseract==0.3.13
reportlab==4.2.5
cloudinary==1.42.2
httpx==0.28.1
python-dotenv==1.0.1
psycopg[binary]==3.2.3
pytest==8.3.4
pytest-asyncio==0.25.0
